patch socket for all py
